pyproject.toml
[build-system]
requires = ["setuptools>=40.8.0", "grpcio-tools>=1.76.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sondera-harness"
version = "0.6.0"
description = "Sondera Harness SDK for Python - Agent governance and policy enforcement"
readme = "README.md"
requires-python = ">=3.12,<3.15"
license = "MIT"
authors = [
    { name = "Sondera AI", email = "sdk@sondera.ai" }
]
keywords = ["ai", "agents", "governance", "policy", "guardrails", "llm", "langchain", "langgraph"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Security",
    "Typing :: Typed",
]
dependencies = [
    "cedar-python>=0.1.1",
    "click>=8.0.0",
    "click-default-group>=1.2.4",
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "httpx>=0.27.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.12.0",
    "textual>=6.11.0",
]

[project.urls]
Homepage = "https://github.com/sondera-ai/harness-sdk-python"
Documentation = "https://docs.sondera.ai"
Repository = "https://github.com/sondera-ai/harness-sdk-python"
Issues = "https://github.com/sondera-ai/harness-sdk-python/issues"
Changelog = "https://github.com/sondera-ai/harness-sdk-python/blob/main/CHANGELOG.md"

[project.scripts]
sondera = "sondera.cli:cli"

[tool.setuptools.package-data]
sondera = ["py.typed"]

[project.optional-dependencies]
adk = ["google-adk>=1.22.0"]
langgraph = ["langchain>=1.2.0", "langgraph>=1.0.5"]
strands = ["strands-agents>=1.21.0"]
all = [
    "google-adk>=1.22.0",
    "langchain>=1.2.0",
    "langgraph>=1.0.5",
    "strands-agents>=1.21.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "textual-dev>=1.8.0",
    "ruff>=0.8.0",
    "pyright",
]
examples = [
    "sondera-examples-adk",
    "sondera-examples-langgraph",
    "sondera-examples-strands",
    "sondera-examples-cedar"
]

[tool.ruff]
line-length = 88
src = ["src", "tests"]
exclude = [
    "src/sondera/proto",      # Generated protobuf files
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
]

ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
# Required for ADK's function schema introspection.
"examples/adk/**" = ["UP006", "UP035", "UP045"]  # Allow deprecated typing syntax (List, Optional, etc.)

[tool.ruff.lint.isort]
known-first-party = ["sondera"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pyright]
exclude = [
    "**/node_modules",
    "**/__pycache__",
    "**/.*",
    "build",
    "dist",
    "src/sondera/proto",  # Generated protobuf files
    "tests",
]

# Define execution environments for main SDK and examples
[[tool.pyright.executionEnvironments]]
root = "src"

[[tool.pyright.executionEnvironments]]
root = "examples/archetypes/src"

[[tool.pyright.executionEnvironments]]
root = "examples/adk/src"
extraPaths = ["src", "examples/archetypes/src"]

[[tool.pyright.executionEnvironments]]
root = "examples/langgraph/src"
extraPaths = ["src", "examples/archetypes/src"]

[[tool.pyright.executionEnvironments]]
root = "examples/strands/src"
extraPaths = ["src", "examples/archetypes/src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
addopts = "-v"

[tool.uv.workspace]
members = ["examples/adk", "examples/langgraph", "examples/strands", "examples/archetypes", "examples/cedar"]

[tool.uv.sources]
sondera-archetypes = { workspace = true }
sondera-examples-adk = { workspace = true }
sondera-examples-langgraph = { workspace = true }
sondera-examples-strands = { workspace = true }
sondera-examples-cedar = { workspace = true }
